backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pillow==10.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
alembic==1.12.1
pytest==7.4.3
httpx==0.25.2
strawberry-graphql==0.211.1
celery==5.3.6
redis==5.0.1
sendgrid==6.10.0
pydantic[email]